api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.1
pydantic==2.7.1
psycopg2-binary==2.9.9
